add forbid warnings

+++ lib/src/deflate/deflate_reader.rs
@@ -164,7 +164,7 @@ impl DeflateParser {
 
             match mode {
                 0 => {
-                    let padding_bits = self.bit_reader.read_padding_bits() as u8;
+                    let padding_bits = self.bit_reader.read_padding_bits();
                     if padding_bits != 0 {
                         return err_exit_code(
                             ExitCode::NonZeroPadding,
